requirements.txt
streamlit
pyautogen
python-dotenv
PyYAML
watchdog 